update google maps api

+++ google_maps_api.py
@@ -1,13 +1,18 @@
 import googlemaps
 import json
-
+from outscraper import ApiClient
+import os.path
+gmaps_key = 'AIzaSyB5T9kvIe2hQZ49w7cpC2B638pdnKpaDTY'
+outscrape_key = 'NGQxMTM4NzBkZmE0NGU0MThhZmRhZDAxMmU0MGIxN2R8NjQ5YWI1ZWM2MQ'
 # Replace 'YOUR_API_KEY' with your actual Google API key
-gmaps = googlemaps.Client(key='AIzaSyB5T9kvIe2hQZ49w7cpC2B638pdnKpaDTY')
+gmaps = googlemaps.Client(key=gmaps_key)
+client = ApiClient(api_key=outscrape_key)
 
+folder = './data/jsons/'
 # Define a function to fetch reviews for a place
 def fetch_reviews(place_id):
     try:
-        place = gmaps.place(place_id=place_id, fields=['name', 'reviews'])
+        place = gmaps.place(place_id=place_id, fields=['name', 'reviews'],)
         if 'reviews' in place['result']:
             return place['result']['reviews']
         else:
@@ -15,6 +20,10 @@ def fetch_reviews(place_id):
     except Exception as e:
         print(f"Error fetching reviews for place {place_id}: {str(e)}")
         return []
+
+def fetch_reviews2(place_id, limit):
+    results = client.google_maps_reviews_v2(query=place_id, reviews_limit=limit, fields=['name', 'reviews'])
+    return results
 
 # Perform a search for hospitals in the U.S.
 search_result = gmaps.places_nearby(
@@ -31,23 +40,29 @@ all_reviews = {
 review_count = 0
 for hospital in search_result.get('results', []):
     place_id = hospital['place_id']
-    reviews = fetch_reviews(place_id)
-    review_count = review_count + len(reviews)
-    location = hospital['geometry']['location']
-    new_hospital = {
-        'hospital_name': hospital['name'],
-        'place_id':place_id,
-        'reviews':reviews,
-        'location':location
-    }
-    all_reviews['hospitals'].append(new_hospital)
-    if review_count >= 1000:
-        break
+    # reviews = fetch_reviews(place_id)
+    output_file_path = folder + place_id + '.json'
+    if not os.path.isfile(output_file_path):
+        reviews = fetch_reviews2(place_id, 20)
+        review_count = review_count + len(reviews)
+        location = hospital['geometry']['location']
+        new_hospital = {
+            'hospital_name': hospital['name'],
+            'place_id':place_id,
+            'reviews':reviews,
+            'location':location
+        }
+        
+        # all_reviews['hospitals'].append(new_hospital)
+        with open(output_file_path,'w') as json_file:
+            json.dump(new_hospital,json_file,indent=4)
+        if review_count >= 100:
+            break
 
-output_file_path = './data/all_reviews.json'
+# output_file_path = './data/all_reviews.json'
 
-with open(output_file_path,'w') as json_file:
-    json.dump(all_reviews,json_file,indent=4)
+# with open(output_file_path,'w') as json_file:
+#     json.dump(new_hospital,json_file,indent=4)
 # Print the most recent 100 reviews
 # for i, review in enumerate(all_reviews[:1000]):
 #     print(f"Review {i + 1}:")
